refactor: extract types and test utils
