+++ src/calc/__tests__/tokenizer.test.ts
@@ -1,4 +1,6 @@
-import { tokenizer, token, TokenType, Operation } from '../tokenizer'
+import { tokenizer } from '../tokenizer'
+import { token } from '../utils'
+import { TokenType, Operation } from '../types'
 
 describe('Tokenizer', () => {
   // Valid cases
